+++ src/main/java/com/domain/redstonetools/features/arguments/serializers/IntegerSerializer.java
@@ -1,4 +1,4 @@
-package com.domain.redstonetools.features.arguments;
+package com.domain.redstonetools.features.arguments.serializers;
 
 import com.mojang.brigadier.arguments.ArgumentType;
 import com.mojang.brigadier.arguments.IntegerArgumentType;
