+++ src/main/java/com/domain/redstonetools/features/toggleable/AirPlaceFeature.java
@@ -3,7 +3,7 @@ package com.domain.redstonetools.features.toggleable;
 import com.domain.redstonetools.features.Feature;
 import com.domain.redstonetools.features.arguments.Argument;
 
-import static com.domain.redstonetools.features.arguments.BoolSerializer.bool;
+import static com.domain.redstonetools.features.arguments.serializers.BoolSerializer.bool;
 
 @Feature(name = "Air Place", description = "Allows you to place blocks in the air.", command = "airplace")
 public class AirPlaceFeature extends ToggleableFeature {
